+++ R/AutoFSelect.R
@@ -48,8 +48,8 @@
 #' # Aggregate performance of outer results
 #' rr$aggregate()
 #'
-#' # Retrieve inner tuning results.
-#' rr$data$learner[[1]]$tuning_result
+#' # Retrieve inner feature selection results
+#' as.data.table(rr)$learner[[1]]$fselect_result
 AutoFSelect = R6Class("AutoFSelect",
   inherit = Learner,
   public = list(
@@ -145,7 +145,7 @@ AutoFSelect = R6Class("AutoFSelect",
 
   active = list(
 
-    #' @field archive ([bbotk::Archive])\cr
+    #' @field archive ([ArchiveFSelect)\cr
     #' Returns [FSelectInstanceSingleCrit] archive.
     archive = function() self$fselect_instance$archive,