+++ R/FSelector.R
@@ -106,6 +106,8 @@ FSelector = R6Class("FSelector",
         bbotk_reflections$optimizer_properties,
         empty.ok = FALSE)
       self$packages = assert_set(packages)
+
+      check_packages_installed(self$packages, msg = sprintf("Package '%%s' required but not installed for FSelector '%s'", format(self)))
     },
 
     #' @description
@@ -129,7 +131,7 @@ FSelector = R6Class("FSelector",
     #' @description
     #' Performs the feature selection on a [FSelectInstanceSingleCrit] or
     #' [FSelectInstanceMultiCrit] until termination.
-    #' The single evaluations will be written into the [bbotk::Archive] that resides in the
+    #' The single evaluations will be written into the [ArchiveFSelect] that resides in the
     #' [FSelectInstanceSingleCrit] / [FSelectInstanceMultiCrit].
     #' The result will be written into the instance object.
     #'