+++ tests/testthat/test_FSelectInstanceMultiCrit.R
@@ -16,12 +16,12 @@ test_that("eval_batch works", {
     x3 = list(TRUE, TRUE), x4 = list(TRUE, TRUE))
 
   z = inst$eval_batch(xdt)
-  expect_named(z, c("regr.mse", "regr.rmse", "resample_result"))
+  expect_named(z, c("regr.mse", "regr.rmse", "uhash"))
   expect_identical(inst$archive$n_evals, 2L)
   expect_data_table(z, nrows = 2L)
 
   z = inst$eval_batch(xdt)
-  expect_named(z, c("regr.mse", "regr.rmse", "resample_result"))
+  expect_named(z, c("regr.mse", "regr.rmse", "uhash"))
   expect_identical(inst$archive$n_evals, 4L)
   expect_data_table(z, nrows = 2L)
 })
@@ -32,17 +32,23 @@ test_that("objective_function works", {
   expect_named(y, c("regr.mse", "regr.rmse"))
 })
 
-test_that("store_resample_results flag works", {
-  inst = TEST_MAKE_INST_2D(store_resample_results = FALSE)
+test_that("store_benchmark_result flag works", {
+  inst = TEST_MAKE_INST_2D(store_benchmark_result = FALSE)
   xdt = data.table(x1 = list(TRUE, FALSE), x2 = list(FALSE, TRUE),
     x3 = list(TRUE, TRUE), x4 = list(TRUE, TRUE))
   inst$eval_batch(xdt)
 
-  expect_true("resample_result" %nin% colnames(inst$archive$data()))
+  expect_true("uhashes" %nin% colnames(inst$archive$data()))
+
+  inst = TEST_MAKE_INST_2D(store_benchmark_result = TRUE)
+  xdt = data.table(x1 = list(TRUE, FALSE), x2 = list(FALSE, TRUE),
+                   x3 = list(TRUE, TRUE), x4 = list(TRUE, TRUE))
+  inst$eval_batch(xdt)
+  expect_r6(inst$archive$benchmark_result, "BenchmarkResult")
 })
 
 test_that("result$features works", {
-  inst = TEST_MAKE_INST_2D(store_resample_results = FALSE)
+  inst = TEST_MAKE_INST_2D(store_benchmark_result = FALSE)
   xdt = data.table(x1 = list(TRUE, FALSE), x2 = list(FALSE, TRUE),
     x3 = list(TRUE, TRUE), x4 = list(TRUE, TRUE))
   ydt = data.table(regr.mse = list(0.1, 0.2),
@@ -51,7 +57,7 @@ test_that("result$features works", {
   expect_list(inst$result_feature_set)
   expect_character(inst$result_feature_set[[1]])
 
-  inst = TEST_MAKE_INST_2D(store_resample_results = FALSE)
+  inst = TEST_MAKE_INST_2D(store_benchmark_result = FALSE)
   xdt = data.table(x1 = list(TRUE), x2 = list(FALSE),
     x3 = list(TRUE), x4 = list(TRUE))
   ydt = data.table(regr.mse = list(0.1),