+++ NEWS.md
@@ -1,3 +1,13 @@
+# mlr3fselect 0.3.0
+
+* `Archive` is `ArchiveFSelect` now which stores the benchmark result in
+  `$benchmark_result`. This change removed the resample results from the archive
+   but they can be still accessed via the benchmark result.
+
+# mlr3fselect 0.2.1
+
+* Warning message if external package for feature selection is not installed.
+
 # mlr3fselect 0.2.0
 
-* First version of mlr3fselect
+* Initial CRAN release.